Update readme to include install information
I had to dig a little bit before I understood that :faker needed to be added to the readme. I also threw in a troubleshooting bullet that showed the error I got (for googling purposes) and a bullet for creating test map templates in Blacksmith. 

Stripping out my Faker out of Blacksmith in favor of yours. It's much better. 

Thanks for a great project!